+++ packages/zenrouter_core/test/mixin/guard_rule_test.dart
@@ -446,7 +446,7 @@ class _MockStackPath implements StackPath<TestRoute> {
   void notifyListeners() {}
 
   @override
-  void clear() {}
+  void clear({RouteTarget? keep}) {}
 
   @override
   void bindStack(List<TestRoute> stack) {}